+++ app/models/concerns/folio/dragonfly_format_validation.rb
@@ -5,13 +5,29 @@ module Folio::DragonflyFormatValidation
 
   class_methods do
     def validate_file_format(formats = %w[jpeg png bmp gif svg tiff])
-      validates_property :format,
-                         of: :file,
-                         in: formats,
-                         message: proc { |actual, model|
-                           I18n.t("activerecord.errors.messages.file_format",
-                                  types: formats.join(", "))
-                         }
+      mime_types = formats.map { |f| "image/#{f}" }
+
+      define_singleton_method :valid_mime_types do
+        mime_types
+      end
+
+      define_singleton_method :valid_mime_types_message do
+        I18n.t("activerecord.errors.messages.file_format",
+               types: formats.join(", "))
+      end
+
+      validate :validate_file_format_via_mime_type
     end
   end
+
+  private
+    def validate_file_format_via_mime_type
+      if mime_type.blank?
+        errors.add(:mime_type, :blank)
+      else
+        if self.class.valid_mime_types.exclude?(mime_type)
+          errors.add(:mime_type, self.class.valid_mime_types_message)
+        end
+      end
+    end
 end